+++ examples/src/main/java/fr/bmartel/speedtest/examples/FixedTimeDownloadExample.java
@@ -29,6 +29,7 @@ import fr.bmartel.speedtest.SpeedTestError;
 import fr.bmartel.speedtest.SpeedTestReport;
 import fr.bmartel.speedtest.SpeedTestSocket;
 
+import java.math.BigDecimal;
 import java.util.Timer;
 import java.util.TimerTask;
 
@@ -86,9 +87,8 @@ public class FixedTimeDownloadExample {
         speedTestSocket.addSpeedTestListener(new ISpeedTestListener() {
 
             @Override
-            public void onDownloadPacketsReceived(final long packetSize, final float transferRateBitPerSeconds, final
-            float
-                    transferRateOctetPerSeconds) {
+            public void onDownloadPacketsReceived(final long packetSize, final BigDecimal transferRateBitPerSeconds,
+                                                  final BigDecimal transferRateOctetPerSeconds) {
                 //called when download is finished
             }
 
@@ -105,9 +105,8 @@ public class FixedTimeDownloadExample {
             }
 
             @Override
-            public void onUploadPacketsReceived(final long packetSize, final float transferRateBitPerSeconds, final
-            float
-                    transferRateOctetPerSeconds) {
+            public void onUploadPacketsReceived(final long packetSize, final BigDecimal transferRateBitPerSeconds,
+                                                final BigDecimal transferRateOctetPerSeconds) {
                 //called when upload is finished
             }
